# server/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0"
  title: "Custom Vision Training 1.2"
host: "southcentralus.api.cognitive.microsoft.com"
basePath: "/customvision/v1.2/Training"
schemes:
- "https"
security:
- apiKeyHeader:
  - "e9b1d79f98294687bf60eef35102d55c"
  - "3a9146cc607a47f7b727b462ccb5aa86"
- apiKeyQuery:
  - "e9b1d79f98294687bf60eef35102d55c"
  - "3a9146cc607a47f7b727b462ccb5aa86"
paths:
  /projects/{projectId}/images:
    post:
      summary: "CreateImagesFromData"
      description: "This API accepts body content as multipart/form-data and application/octet-stream.\
        \ When using multipart\r\nmultiple image files can be sent at once, with a\
        \ maximum of 64 files"
      operationId: "createImagesFromData"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "tagIds"
        in: "query"
        description: "The tags ids with which to tag each image. Limited to 20"
        required: false
        type: "string"
      - name: "imageData"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ImageCreateSummary"
      x-swagger-router-controller: "Default"
    delete:
      summary: "DeleteImages"
      description: "Delete images from the set of training images"
      operationId: "deleteImages"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "imageIds"
        in: "query"
        description: "Ids of the images to be deleted. Limted to 256 images per batch"
        required: true
        type: "string"
      responses:
        204:
          description: "No Content"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/images/files:
    post:
      summary: "CreateImagesFromFiles"
      description: "Add the provided batch of images to the set of training images"
      operationId: "createImagesFromFiles"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - in: "body"
        name: "imageFileCreateBatch"
        description: "The batch of image files to add. Limited to 64 images and 20\
          \ tags per batch"
        required: false
        schema:
          $ref: "#/definitions/ImageFileCreateBatch"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ImageCreateSummary"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/images/predictions:
    post:
      summary: "CreateImagesFromPredictions"
      description: "Add the specified predicted images to the set of training images"
      operationId: "createImagesFromPredictions"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - in: "body"
        name: "imageIdCreateBatch"
        description: "Image and tag ids. Limted to 64 images and 20 tags per batch"
        required: false
        schema:
          $ref: "#/definitions/ImageIdCreateBatch"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ImageCreateSummary"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/images/urls:
    post:
      summary: "CreateImagesFromUrls"
      description: "Add the provided images urls to the set of training images"
      operationId: "createImagesFromUrls"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - in: "body"
        name: "imageUrlCreateBatch"
        description: "Image urls and tag ids. Limited to 64 images and 20 tags per\
          \ batch"
        required: false
        schema:
          $ref: "#/definitions/ImageUrlCreateBatch"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ImageCreateSummary"
      x-swagger-router-controller: "Default"
  /projects:
    get:
      summary: "GetProjects"
      description: "Get your projects"
      operationId: "getProjects"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ProjectArray"
      x-swagger-router-controller: "Default"
    post:
      summary: "CreateProject"
      description: "Create a project"
      operationId: "createProject"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "name"
        in: "query"
        description: "Name of the project"
        required: true
        type: "string"
      - name: "description"
        in: "query"
        description: "The description of the project"
        required: false
        type: "string"
      - name: "domainId"
        in: "query"
        description: "Format - uuid. The id of the domain to use for this project.\
          \ Defaults to General"
        required: false
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Project"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/tags:
    get:
      summary: "GetTags"
      description: "Get the tags for a given project and iteration"
      operationId: "getTags"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "iterationId"
        in: "query"
        description: "Format - uuid. The iteration id. Defaults to workspace"
        required: false
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/TagList"
      x-swagger-router-controller: "Default"
    post:
      summary: "CreateTag"
      description: "Create a tag for the project"
      operationId: "createTag"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "name"
        in: "query"
        description: "The tag name"
        required: true
        type: "string"
      - name: "description"
        in: "query"
        description: "Optional description for the tag"
        required: false
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Tag"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/images/tags:
    post:
      summary: "PostImageTags"
      description: "Associate a set of images with a set of tags"
      operationId: "postImageTags"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - in: "body"
        name: "imageTagCreateBatch"
        description: "Batch of image tags. Limited to 128 tags per batch"
        required: false
        schema:
          $ref: "#/definitions/ImageTagCreateBatch"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ImageTagCreateSummary"
      x-swagger-router-controller: "Default"
    delete:
      summary: "DeleteImageTags"
      description: "Remove a set of tags from a set of images"
      operationId: "deleteImageTags"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "imageIds"
        in: "query"
        description: "Image ids. Limited to 64 images"
        required: true
        type: "string"
      - name: "tagIds"
        in: "query"
        description: "Tags to be deleted from the specified images. Limted to 20 tags"
        required: true
        type: "string"
      responses:
        204:
          description: "No Content"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/iterations/{iterationId}:
    get:
      summary: "GetIteration"
      description: "Get a specific iteration"
      operationId: "getIteration"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The id of the project the iteration belongs to"
        required: true
        type: "string"
      - name: "iterationId"
        in: "path"
        description: "Format - uuid. The id of the iteration to get"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Iteration"
      x-swagger-router-controller: "Default"
    delete:
      summary: "DeleteIteration"
      description: "Delete a specific iteration of a project"
      operationId: "deleteIteration"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "iterationId"
        in: "path"
        description: "Format - uuid. The iteration id"
        required: true
        type: "string"
      responses:
        204:
          description: "No Content"
      x-swagger-router-controller: "Default"
    patch:
      summary: "UpdateIteration"
      description: "Update a specific iteration"
      operationId: "updateIteration"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. Project id"
        required: true
        type: "string"
      - name: "iterationId"
        in: "path"
        description: "Format - uuid. Iteration id"
        required: true
        type: "string"
      - in: "body"
        name: "iteration"
        description: "The updated iteration model"
        required: false
        schema:
          $ref: "#/definitions/Iteration"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Iteration"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/predictions:
    delete:
      summary: "DeletePrediction"
      description: "Delete a set of predicted images and their associated prediction\
        \ results"
      operationId: "deletePrediction"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "ids"
        in: "query"
        description: "The prediction ids. Limited to 64"
        required: true
        type: "string"
      responses:
        204:
          description: "No Content"
      x-swagger-router-controller: "Default"
  /projects/{projectId}:
    get:
      summary: "GetProject"
      description: "Get a specific project"
      operationId: "getProject"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The id of the project to get"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Project"
      x-swagger-router-controller: "Default"
    delete:
      summary: "DeleteProject"
      description: "Delete a specific project"
      operationId: "deleteProject"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      responses:
        204:
          description: "No Content"
      x-swagger-router-controller: "Default"
    patch:
      summary: "UpdateProject"
      description: "Update a specific project"
      operationId: "updateProject"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The id of the project to update"
        required: true
        type: "string"
      - in: "body"
        name: "project"
        description: "The updated project model"
        required: false
        schema:
          $ref: "#/definitions/Project"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Project"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/tags/{tagId}:
    get:
      summary: "GetTag"
      description: "Get information about a specific tag"
      operationId: "getTag"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project this tag belongs to"
        required: true
        type: "string"
      - name: "tagId"
        in: "path"
        description: "Format - uuid. The tag id"
        required: true
        type: "string"
      - name: "iterationId"
        in: "query"
        description: "Format - uuid. The iteration to retrieve this tag from. Optional,\
          \ defaults to current training set"
        required: false
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Tag"
      x-swagger-router-controller: "Default"
    delete:
      summary: "DeleteTag"
      description: "Delete a tag from the project"
      operationId: "deleteTag"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "tagId"
        in: "path"
        description: "Format - uuid. Id of the tag to be deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "No Content"
      x-swagger-router-controller: "Default"
    patch:
      summary: "UpdateTag"
      description: "Update a tag"
      operationId: "updateTag"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "tagId"
        in: "path"
        description: "Format - uuid. The id of the target tag"
        required: true
        type: "string"
      - in: "body"
        name: "tag"
        description: "The updated tag model"
        required: false
        schema:
          $ref: "#/definitions/Tag"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Tag"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/iterations/{iterationId}/export:
    get:
      summary: "GetExports"
      description: "Get the list of exports for a specific iteration"
      operationId: "getExports"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "iterationId"
        in: "path"
        description: "Format - uuid. The iteration id"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ExportArray"
      x-swagger-router-controller: "Default"
    post:
      summary: "ExportIteration"
      description: "Export a trained iteration"
      operationId: "exportIteration"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      - name: "iterationId"
        in: "path"
        description: "Format - uuid. The iteration id"
        required: true
        type: "string"
      - name: "platform"
        in: "query"
        description: "The target platform (coreml or tensorflow)"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Export"
      x-swagger-router-controller: "Default"
...
  /domains/{domainId}:
    get:
      summary: "GetDomain"
      description: "Get information about a specific domain"
      operationId: "getDomain"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "domainId"
        in: "path"
        description: "Format - uuid. The id of the domain to get information about"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Domain"
      x-swagger-router-controller: "Default"
...
  /projects/{projectId}/iterations:
    get:
      summary: "GetIterations"
      description: "Get iterations for the project"
      operationId: "getIterations"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "projectId"
        in: "path"
        description: "Format - uuid. The project id"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/IterationArray"
      x-swagger-router-controller: "Default"
...
securityDefinitions:
  apiKeyHeader:
    type: "apiKey"
    name: "Training-key"
    in: "header"
  apiKeyQuery:
    type: "apiKey"
    name: "Training-key"
    in: "query"
definitions:
...
  ImageFileCreateBatch:
    type: "object"
    properties:
      Images:
        type: "array"
        items:
          $ref: "#/definitions/ImageFileCreateEntry"
        x-nullable: true
      TagIds:
        type: "array"
        items:
          type: "string"
          format: "uuid"
          x-nullable: false
        x-nullable: true
    example:
      TagIds:
      - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      Images:
      - Contents: "Contents"
        TagIds:
        - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        Name: "Name"
      - Contents: "Contents"
        TagIds:
        - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        Name: "Name"
    x-nullable: true
...
  ImageUrlCreateBatch:
    type: "object"
    properties:
      Images:
        type: "array"
        items:
          $ref: "#/definitions/ImageUrlCreateEntry"
        x-nullable: true
      TagIds:
        type: "array"
        items:
          type: "string"
          format: "uuid"
          x-nullable: false
        x-nullable: true
    example:
      TagIds:
      - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      Images:
      - TagIds:
        - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        Url: "Url"
      - TagIds:
        - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        - "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        Url: "Url"
    x-nullable: true
...